Fix issue where storage directory could not be parsed as a string
Using PathLib vs os with the latest django exposed an issue where the BASE_DIR configuration is not an os string path, resulting in an error when fetching images from the media url directory. The fix is to stick with the PathLib syntax but convert the path to a string.